patientFullName -> getPatientFullName and move from format to misc utils b/c doesn't do any formatting
