+++ src/utils/misc.js
@@ -1,0 +1,34 @@
+/*
+ * == BSD2 LICENSE ==
+ * Copyright (c) 2017, Tidepool Project
+ *
+ * This program is free software; you can redistribute it and/or modify it under
+ * the terms of the associated License, which is identical to the BSD 2-Clause
+ * License as published by the Open Source Initiative at opensource.org.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the License for more details.
+ *
+ * You should have received a copy of the License along with this program; if
+ * not, you can obtain one from Tidepool Project at tidepool.org.
+ * == BSD2 LICENSE ==
+ */
+
+import _ from 'lodash';
+
+/**
+ * getPatientFullName
+ * @param {Object} patient - Tidepool patient object containing profile
+ *
+ * @return {String} PwD's full name
+ */
+export function getPatientFullName(patient) {
+  const profile = _.get(patient, ['profile'], {});
+  const patientInfo = profile.patient || {};
+
+  if (patientInfo.isOtherPerson) {
+    return patientInfo.fullName;
+  }
+  return profile.fullName;
+}
